update ratatui and support filled tris
