+++ braillix_ratatui/Cargo.toml
@@ -14,4 +14,5 @@ repository.workspace = true
 [dependencies]
 braillix = { path = "../braillix", version = "0.2.0" }
 crossterm = "0.28.1"
-ratatui = "0.29.0"
+ratatui-core = "0.1.0"
+ratatui = "0.30.0"
